还款新增结束债权操作

+++ src/main/java/com/gofobao/framework/tender/entity/VirtualTender.java
@@ -22,13 +22,13 @@ public class VirtualTender {
     @Id
     @Column(name = "id", nullable = false)
     @GeneratedValue
-    private Integer id;
+    private Long id;
     @Basic
     @Column(name = "user_id", nullable = false)
     private Long userId;
     @Basic
     @Column(name = "borrow_id", nullable = false)
-    private Integer borrowId;
+    private Long borrowId;
     @Basic
     @Column(name = "status", nullable = false)
     private Integer status;
